pyproject.toml
[project]
name = "project-file"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "easyocr>=1.7.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "rapidfuzz>=3.14.3",
    "streamlit>=1.55.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.3.0",
    "ultralytics>=8.4.21",
]
